requirements.txt

pynput==1.7.6
pillow==10.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
